[IMP] mail: composer placeholder
Composer placeholder must adapt to the number of channel members
